Stop flicker.
